# spec/openapi.yaml
openapi: "3.1.0"
info:
  title: "标题：ArchForge接口文档"
  description: "描述：文档说明"
  contact:
    name: "ArchForge"
  version: "版本号:1.0.0"
  license:
    name: "Apache-2.0"
    identifier: "Apache-2.0"
servers:
- url: "/"
security:
- bearer: []
...
paths:
...
  /admin/system/dict/item/{id}:
    put:
      tags:
      - "字典配置"
      summary: "更新字典项"
      operationId: "updateItem"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictItemRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
    delete:
      tags:
      - "字典配置"
      summary: "删除字典项"
      operationId: "deleteItem"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/system/dict/type:
    post:
      tags:
      - "字典配置"
      summary: "字典类型分页列表"
      operationId: "listTypes"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictTypeQueryRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPageResponseDictTypeResponse"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/system/dict/type/create:
    post:
      tags:
      - "字典配置"
      summary: "创建字典类型"
      operationId: "createType"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictTypeCreateRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "integer"
                format: "int64"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/system/dict/type/{typeId}/item:
    post:
      tags:
      - "字典配置"
      summary: "创建字典项"
      operationId: "createItem"
      parameters:
      - name: "typeId"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictItemRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "integer"
                format: "int64"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/system/dict/type/{typeKey}:
    get:
      tags:
      - "字典配置"
      summary: "根据字典编码查询详情及全部项"
      operationId: "getTypeByCode"
      parameters:
      - name: "typeKey"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DictTypeResponse"
        "4XX":
          $ref: "#/components/responses/Error"
    put:
      tags:
      - "字典配置"
      summary: "更新字典类型"
      operationId: "updateType"
      parameters:
      - name: "typeKey"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DictTypeUpdateRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
    delete:
      tags:
      - "字典配置"
      summary: "删除字典类型"
      operationId: "deleteType"
      parameters:
      - name: "typeKey"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
...
components:
  schemas:
...
    AdminPageResponseDictTypeResponse:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            $ref: "#/components/schemas/DictTypeResponse"
        total:
          type: "integer"
          format: "int64"
        pageSize:
          type: "integer"
          format: "int32"
        currentPage:
          type: "integer"
          format: "int32"
...
    DictItemRequest:
      required:
      - "itemCode"
      - "itemLabel"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        itemCode:
          minLength: 1
          type: "string"
        itemLabel:
          minLength: 1
          type: "string"
        sort:
          type: "integer"
          format: "int32"
        status:
          type: "integer"
          format: "int32"
    DictItemResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        itemCode:
          type: "string"
        itemLabel:
          type: "string"
        sort:
          type: "integer"
          format: "int32"
        status:
          type: "integer"
          format: "int32"
        dictTypeId:
          type: "integer"
          format: "int64"
    DictTypeCreateRequest:
      required:
      - "dictCode"
      - "dictName"
      type: "object"
      properties:
        dictCode:
          minLength: 1
          type: "string"
        dictName:
          minLength: 1
          type: "string"
        description:
          type: "string"
        status:
          type: "integer"
          format: "int32"
        sort:
          type: "integer"
          format: "int32"
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/DictItemRequest"
    DictTypeQueryRequest:
      type: "object"
      properties:
        currentPage:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
        keyword:
          type: "string"
    DictTypeResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        dictCode:
          type: "string"
        dictName:
          type: "string"
        description:
          type: "string"
        status:
          type: "integer"
          format: "int32"
        sort:
          type: "integer"
          format: "int32"
        createTime:
          type: "string"
          format: "date-time"
        updateTime:
          type: "string"
          format: "date-time"
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/DictItemResponse"
    DictTypeUpdateRequest:
      required:
      - "dictName"
      type: "object"
      properties:
        dictName:
          minLength: 1
          type: "string"
        description:
          type: "string"
        status:
          type: "integer"
          format: "int32"
        sort:
          type: "integer"
          format: "int32"
...
  securitySchemes:
    bearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  responses:
    Error:
      description: "Error response (RFC 9457 ProblemDetail)"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              type:
                type: "string"
              title:
                type: "string"
              status:
                type: "integer"
              detail:
                type: "string"
              instance:
                type: "string"